Initial broadcom port for Raspberry Pi
This targets the 64-bit CPU Raspberry Pis. The BCM2711 on the Pi 4
and the BCM2837 on the Pi 3 and Zero 2W. There are 64-bit fixes
outside of the ports directory for it.

There are a couple other cleanups that were incidental:
* Use const mcu_pin_obj_t instead of omitting the const. The structs
  themselves are const because they are in ROM.
* Use PTR <-> OBJ conversions in more places. They were found when
  mp_obj_t was set to an integer type rather than pointer.
* Optimize submodule checkout because the Pi submodules are heavy
  and unnecessary for the vast majority of builds.

Fixes #4314

+++ shared-module/displayio/Bitmap.h
@@ -37,8 +37,8 @@ typedef struct {
     mp_obj_base_t base;
     uint16_t width;
     uint16_t height;
-    size_t *data;
-    uint16_t stride; // size_t's
+    uint32_t *data;
+    uint16_t stride; // uint32_t's
     uint8_t bits_per_value;
     uint8_t x_shift;
     size_t x_mask;
